fix some minor association problems

+++ db/migrate/20160922211748_create_children.rb
@@ -4,7 +4,7 @@ class CreateChildren < ActiveRecord::Migration[5.0]
       t.string :name, null: false
       t.string :password_digest, null: false
       t.references :parent, null: false
-      t.float :undeposited_funds, default: 0
+      t.integer :undeposited_funds, default: 0
       t.timestamps
     end
   end
